tests/wpt/tests/css/css-view-transitions/navigation/with-types/at-rule-opt-in-auto-with-types.html

<!DOCTYPE html>
<title>View Transitions: @view-transition opt in with types</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@view-transition {
  navigation: auto;
  types: mytype;
}
</style>
<script>
const params = new URLSearchParams(location.search);
const mode = params.get("mode") || "test";

switch (mode) {
case "test":
  promise_test(async t => {
    const result = await new Promise(resolve => {
      window.did_reveal = msg => { resolve(msg) };
      const popup = window.open("?mode=old");
      t.add_cleanup(() => popup.close());
    });
    assert_equals(result, "pass");
  });
  break;
case "old":
  onload = () => requestAnimationFrame(() => requestAnimationFrame(() => {
    location.replace("?mode=new");
  }));
  break;
case "new":
  onpagereveal = e => {
    if (!e.viewTransition) {
      window.opener.did_reveal("fail: no viewTransition");
      return;
    }
    if (!e.viewTransition.types.has("mytype")) {
      window.opener.did_reveal("fail: missing mytype");
      return;
    }
    window.opener.did_reveal("pass");
  };
}
</script>
